pyproject.toml
[tool.poetry]
name = "ticketscloud"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.3"
pydantic-settings = "^2.2.1"
motor = "^3.3.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"
mypy = "^1.8.0"
motor-types = "^1.0.0b4"
pytest-aiohttp = "^1.0.5"

[tool.ruff]
line-length = 100
target-version = "py311"

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".pytest_cache",
    "Makefile",
]

[tool.ruff.lint]
# list of rules - https://docs.astral.sh/ruff/rules/
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "I",  # isort
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
warn_no_return = false
check_untyped_defs = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
